Create DBH_Workgiver.xml
Added the MO DBH building to the DoBillsBurnPit WorkGiverDef so pawns can do the bills for it. Previously pawns would not cremate bodies using this building. 

There are other ways to do this such as creating a new WorkGiverDef but this approach seems to be more simplistic and still functional.